zero-domestic-control/pyproject.toml
[project]
name = "zero-domestic-control"
description = ""
authors = [
    { name = "Boudewijn van Groos", email = "boudewijn.vangroos@foundationzero.org>" },
    { name = "Floris Smit", email = "floris.smit@foundationzero.org" }
]
readme = "README.md"
packages = [{ include = "io_processing", from = "src" }]
requires-python = ">= 3.13.1,<4.0"
dynamic = ["dependencies", "version"]

[tool.poetry]
version = "0.0.0"

[tool.poetry.dependencies]
starlette-graphene3 = "^0.6.0"
strawberry-graphql = {extras = ["debug-server", "fastapi"], version = "^0.256.1"}
pymodbustcp = "^0.3.0"
fastapi = {extras = ["standard"], version = "^0.115.6"}
aiomqtt = "^2.3.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
pyjwt = "^2.10.1"
pydantic-settings = "^2.7.1"
homeassistant-api = "^5.0.0"


[tool.poetry.requires-plugins]
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
ruff = "^0.9.5"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.1"
httpx = "^0.28.1"
pytest-timeout = "^2.3.1"
pytest-mock = "^3.14.0"

[tool.mypy]
plugins = "pydantic.mypy,strawberry.ext.mypy_plugin"
mypy_path = "./typings"

[tool.poetry-dynamic-versioning]
enable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

[tool.ruff]
extend-exclude = ["hass"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

